translate some code to english
