--- src/components/locais.js
@@ -1,11 +1,0 @@
-import Place from "./place";
-
-export default function Locais(props) {
-  return (
-    <section>
-      {props.locais.map((local) => {
-        return <Place local={local} key={local._id}></Place>;
-      })}
-    </section>
-  );
-}
